mrview view tool: fix signed warning in assert()

+++ src/gui/mrview/tool/view.cpp
@@ -649,7 +649,7 @@ namespace MR
         {
           if (window().image()) {
             const auto& image (window().image()->image);
-            assert (image.ndim() == volume_index_layout->count()+3);
+            assert (image.ndim() == size_t(volume_index_layout->count()+3));
 
             for (int i = 0; i < volume_index_layout->count(); ++i) {
               auto* box = dynamic_cast<SpinBox*> (volume_index_layout->itemAt(i)->widget());
